Customers and Items done

+++ src/main/java/com/qa/project/persistence/domain/Items.java
@@ -3,14 +3,14 @@ package com.qa.project.persistence.domain;
 public class Items {
 	private Long itemId;
 	private String itemName;
-	private int value;
+	private Long value;
 	
-	public Items(String items, int cost) {
+	public Items(String items, Long cost) {
 		this.itemName = items;
 		this.value = cost;
 	}
 	
-	public Items(Long itemId, String items, int cost) {
+	public Items(Long itemId, String items, Long cost) {
 		this.itemId = itemId;
 		this.itemName = items;
 		this.value = cost;
@@ -32,11 +32,11 @@ public class Items {
 		this.itemName = itemName;
 	}
 
-	public int getValue() {
+	public Long getValue() {
 		return value;
 	}
 
-	public void setValue(int value) {
+	public void setValue(Long value) {
 		this.value = value;
 	}
